double -->float for retained parameters

+++ SOC_Photon/src/Coulombs.cpp
@@ -34,7 +34,7 @@ extern CommandPars cp;
 
 // class Coulombs
 Coulombs::Coulombs() {}
-Coulombs::Coulombs(double *rp_delta_q, double *rp_t_last, const double q_cap_rated, const double t_rated, const double t_rlim)
+Coulombs::Coulombs(double *rp_delta_q, float *rp_t_last, const double q_cap_rated, const double t_rated, const double t_rlim)
   : rp_delta_q_(rp_delta_q), rp_t_last_(rp_t_last), q_cap_rated_(q_cap_rated), q_cap_rated_scaled_(q_cap_rated), t_rated_(t_rated), t_rlim_(0.017),
   soc_min_(0)
   {
